+++ archiver/compressor.h
@@ -17,15 +17,15 @@ public:
 
 private:
     void EncodeFileName();
-    void MakeCanonicalHuffmanCode(const std::vector<std::pair<size_t, size_t>>& codes,
-                                  std::unordered_map<size_t, size_t>& cnt_len_code);
-    void WriteCodeTableToFile(size_t max_symbol_code_size,
-                              std::unordered_map<size_t, size_t>& cnt_len_code);
-    BinaryTrie BuildTrie(std::unordered_map<size_t, size_t>& cnt_bytes);
-    std::unordered_map<size_t, size_t> BuildBytesFrequencyMap();
+    void MakeCanonicalHuffmanCode(const std::vector<std::pair<uint16_t, uint16_t>>& codes,
+                                  std::unordered_map<uint16_t, uint16_t>& cnt_len_code);
+    void WriteCodeTableToFile(uint16_t max_symbol_code_size,
+                              std::unordered_map<uint16_t, uint16_t>& cnt_len_code);
+    BinaryTrie BuildTrie(std::unordered_map<uint16_t, size_t>& cnt_bytes);
+    std::unordered_map<uint16_t, size_t> BuildBytesFrequencyMap();
     void EncodeFileAndWriteToArchive();
 
-    using CodeTable = std::unordered_map<size_t, HuffmanCode>;
+    using CodeTable = std::unordered_map<uint16_t, HuffmanCode>;
     CodeTable code_table_;
 
     size_t files_added_ = 0;
